+++ totalRP3/core/impl/locale/locale_enUS.lua
@@ -983,12 +983,56 @@ Your profiles, companions profiles and settings will be temporarily stashed away
 |cff00ff00Use the same command again (|cff999999/trp3 stash|cff00ff00) to restore your data.|r]],
 		OPTION_ENABLED_TOAST = "Option enabled",
 		OPTION_DISABLED_TOAST = "Option disabled",
-		WHATS_NEW_14 = [[
-## 1.2.9.2 - 2017-09-14
+		WHATS_NEW_16 = [[
+## 1.2.11 - 2017-11-09
+
+### Added
+
+- Added support for the profile downloading indicator from and to the XRP add-on.
 
 ### Fixed
 
-- Implemented a temporary workaround for the issue introduced by patch 7.3 in the game's code that cause the misalignment of the text cursor in multi lines text fields. Note that this workaround is not perfect, and result in empty lines visible at the bottom of the text, but at least the cursor is actually on the right line. But Blizzard is aware of the issue and is working on fixing it properly on their end.
+- Fixed an error when trying to whitelist a profile that has been flagged as containing mature content when the profile hasn't been entirely downloaded yet ([ticket #133](https://wow.curseforge.com/projects/total-rp-3/issues/133)).
+- Fixed an issue allowing the user to send empty NPC messages when using the Enter key ([ticket #124](https://wow.curseforge.com/projects/total-rp-3/issues/124)).
+- Fixed an error when targeting battle pets that are participating in a pet battle ([ticket #96](https://wow.curseforge.com/projects/total-rp-3/issues/96)).
+- Fixed an issue where if you used a single space character for your class (like to indicate you have none) it would be considered as empty and your character's real class would be used instead ([ticket #103](https://wow.curseforge.com/projects/total-rp-3/issues/103)).
+- Fixed an issue where players with custom RP status from other add-ons sent via the Mary Sue Protocol would be shown as Out Of Character.
+
+### Removed
+
+- Removed workaround for the text box issue introduced in patch 7.3 as this issue has been fixed in patch 7.3.2.
+
+]],
+		WHATS_NEW_16_1 = [[
+## 1.2.11.1 - 2017-12-08
+
+### Fixed
+
+- Fixed an issue where the Mary Sue Protocol downloading indicator would get stuck for Total RP 3 profiles.
+
+]],
+		WHATS_NEW_16_2 = [[
+## 1.2.11.2 - 2017-12-26
+
+### Fixed
+
+- Fixed a Lua overflow error with the ChatThrottleLib that could occur in rare cases.
+- Fixed an issue that would cause the tooltip to reload all the data too frequently.
+- Fixed an issue that could cause a larger than usual amount of Unknown profiles to be listed in the Directory.
+
+### Removed
+
+- Removed the downloading progression indicator in the tooltip for now as it was the cause of some of these issues. It will be brought back later with a better implementation.
+
+]],
+		WHATS_NEW_16_3 = [[
+## 1.2.11.3 - 2018-01-02
+
+Happy new year! The Total RP 3 team wishes you the best for 2018.
+
+### Updated
+
+- Updated list of Patreon supporters inside the add-on for the month of December.
 
 ]],
 		MORE_MODULES_2 = [[{h2:c}Optional modules{/h2}
@@ -1020,6 +1064,9 @@ The Kui |cff9966ffNameplates|r module adds several Total RP 3 customizations to 
 - {twitter*Solanya_*Solanya} (Community Manager)
 
 {h2}{icon:THUMBUP:20} Acknowledgements{/h2}
+{col:ffffff}Ellypse's {/col}{link*https://www.patreon.com/ellypse*Patreon} {col:ffffff}supporters:{/col}
+%s
+
 {col:ffffff}Our pre-alpha QA team:{/col}
 - Erzan
 - Calian