+++ totalRP3/modules/register/main/register_tooltip.lua
@@ -52,6 +52,7 @@ local lightenColorUntilItIsReadable = Utils.color.lightenColorUntilItIsReadable;
 local isPlayerIC;
 local unitIDIsFilteredForMatureContent;
 local crop = Utils.str.crop;
+local IsAltKeyDown = IsAltKeyDown;
 
 -- ICONS
 local AFK_ICON = "|TInterface\\FriendsFrame\\StatusIcon-Away:15:15|t";