+++ totalRP3/resources/imageList.lua
@@ -283,6 +283,41 @@ local IMAGES = {
 		height = 128
 	},
 	{
+		url = "Interface\\Calendar\\Holidays\\Calendar_Brawl",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\Calendar\\Holidays\\Calendar_DanceDay",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\Calendar\\Holidays\\Calendar_MoonkinFestival",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\Calendar\\Holidays\\Calendar_RunningoftheGnomesEnd",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\Calendar\\Holidays\\Calendar_TombofSargerasStart",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\Calendar\\Holidays\\Calendar_TransmogPopularityContestStart",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\Calendar\\Holidays\\Calendar_WeekendBlackTempleStart",
+		width = 128,
+		height = 128
+	},
+	{
 		url = "Interface\\Challenges\\ChallengeMode_Medal_Bronze",
 		width = 128,
 		height = 128
@@ -326,7 +361,68 @@ local IMAGES = {
 		url = "Interface\\Challenges\\challenges-silver",
 		width = 256,
 		height = 256
-	}, {
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomDeathKnight",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomDemonHunter",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomDruid",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomHunter",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomMage",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomMonk",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomPaladin",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomPriest",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomRogue",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomShaman",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomWarlock",
+		width = 512,
+		height = 512
+	},
+	{
+		url = "Interface\\DRESSUPFRAME\\DressingRoomWarrior",
+		width = 512,
+		height = 512
+	},
+	{
 		url = "Interface\\FlavorImages\\BloodElfLogo-small",
 		width = 256,
 		height = 256
@@ -981,6 +1077,38 @@ local IMAGES = {
 		width = 256,
 		height = 256
 	}, {
+		url = "Interface\\LFGFRAME\\LFGIcon-CathedralOfEternalNight",
+		width = 256,
+		height = 256
+	}, {
+		url = "Interface\\LFGFRAME\\LFGIcon-TombOfSargerasChamberOfTheAvatar",
+		width = 256,
+		height = 256
+	}, {
+		url = "Interface\\LFGFRAME\\LFGIcon-TombOfSargerasDeceiversFall",
+		width = 256,
+		height = 256
+	}, {
+		url = "Interface\\LFGFRAME\\LFGIcon-TombOfSargerasTheGatesOfHell",
+		width = 256,
+		height = 256
+	}, {
+		url = "Interface\\LFGFRAME\\LFGIcon-TombOfSargerasWailingHalls",
+		width = 256,
+		height = 256
+	}, {
+		url = "Interface\\LFGFRAME\\LFGIcon-AntorusWing1",
+		width = 256,
+		height = 256
+	}, {
+		url = "Interface\\LFGFRAME\\LFGIcon-Argus",
+		width = 256,
+		height = 256
+	}, {
+		url = "Interface\\LFGFRAME\\LFGIcon-SeatoftheTriumvirate",
+		width = 256,
+		height = 256
+	}, {
 		url = "Interface\\PETBATTLES\\Weather-ArcaneStorm",
 		width = 512,
 		height = 128
@@ -1244,6 +1372,606 @@ local IMAGES = {
 		url = "Interface\\PVPFrame\\Icons\\prestige-icon-4-5",
 		width = 128,
 		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-5-1",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-5-2",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-5-3",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-5-4",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-6-1",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-6-2",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-6-3",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-6-4",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-7-1",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-7-2",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-7-3",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-7-4",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-6-4",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-8-1",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-8-2",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-8-3",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\prestige-icon-8-4",
+		width = 128,
+		height = 128
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-1",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-2",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-3",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-4",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-5",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-6",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-7",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-8",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-9",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-10",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-11",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-12",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-13",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-14",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-15",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-16",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-17",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-18",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-19",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-20",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-21",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-22",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-23",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-24",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-25",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-26",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-27",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-28",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-29",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-30",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-31",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-32",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-33",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-34",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-35",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-36",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-37",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-38",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-39",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-40",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-41",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-42",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-43",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-44",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-45",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-46",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-47",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-48",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-49",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-50",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-51",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-52",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-53",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-54",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-55",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-56",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-57",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-58",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-59",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-60",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-61",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-62",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-63",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-64",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-65",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-66",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-67",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-68",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-69",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-70",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-71",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-72",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-73",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-74",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-75",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-76",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-77",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-78",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-79",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-80",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-81",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-82",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-83",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-84",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-85",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-86",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-87",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-88",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-89",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-90",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-91",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-92",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-93",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-94",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-95",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-96",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-97",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-98",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-99",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-100",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-101",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-102",
+		width = 64,
+		height = 64
+	},
+	{
+		url = "Interface\\PVPFrame\\Icons\\PVP-Banner-Emblem-103",
+		width = 64,
+		height = 64
 	},
 	{
 		url = "Interface\\QuestionFrame\\answer-ArmoryAlliance",
@@ -1682,6 +2410,46 @@ local IMAGES = {
 	},
 	{
 		url = "Interface\\QuestionFrame\\answer-WarriorProtection",
+		width = 256,
+		height = 128,
+	},
+	{
+		url = "Interface\\QuestionFrame\\answer-ChromieScenario-Chromie",
+		width = 256,
+		height = 128,
+	},
+	{
+		url = "Interface\\QuestionFrame\\answer-ChromieScenario-Drake",
+		width = 256,
+		height = 128,
+	},
+	{
+		url = "Interface\\QuestionFrame\\answer-ChromieScenario-Gold",
+		width = 256,
+		height = 128,
+	},
+	{
+		url = "Interface\\QuestionFrame\\answer-ChromieScenario-Hourglass",
+		width = 256,
+		height = 128,
+	},
+	{
+		url = "Interface\\QuestionFrame\\answer-VindicaarCrystal-FelHeartofArgus",
+		width = 256,
+		height = 128,
+	},
+	{
+		url = "Interface\\QuestionFrame\\answer-VindicaarCrystal-LightforgedWarframe",
+		width = 256,
+		height = 128,
+	},
+	{
+		url = "Interface\\QuestionFrame\\answer-VindicaarCrystal-LightsJudgment",
+		width = 256,
+		height = 128,
+	},
+	{
+		url = "Interface\\QuestionFrame\\answer-VindicaarCrystal-ShroudofArcaneEchoes",
 		width = 256,
 		height = 128,
 	},